avoid using definition function into definition package
