+++ safe/gui/tools/wizard/wizard_dialog.py
@@ -39,7 +39,7 @@ from safe.gui.tools.wizard.wizard_strings import (
     category_question_aggregation)
 from safe.gui.tools.wizard.wizard_utils import layer_description_html
 
-from safe.utilities.settings import set_inasafe_default_value_qsetting
+from safe.utilities.default_values import set_inasafe_default_value_qsetting
 from safe.utilities.gis import (
     is_raster_layer,
     is_point_layer,
