+++ safe/utilities/settings.py
@@ -2,8 +2,6 @@
 """Helper function for InaSAFE settings."""
 
 from PyQt4.QtCore import QSettings
-from safe.definitions import GLOBAL, zero_default_value
-from safe.definitions.utilities import definition
 from safe.utilities.i18n import tr
 
 __copyright__ = "Copyright 2016, The InaSAFE Project"
@@ -133,60 +131,3 @@ def delete_setting(key):
     """
     settings = QSettings()
     settings.remove('%s/%s' % (APPLICATION_NAME, key))
-
-
-def set_inasafe_default_value_qsetting(
-        qsetting, category, inasafe_field_key, value):
-    """Helper method to set inasafe default value to qsetting.
-
-    :param qsetting: QSettings
-    :type qsetting: QSettings
-
-    :param category: Category of the default value. It can be global or
-        recent. Global means the global setting for default value. Recent
-        means the last set custom for default value from the user.
-    :type category: str
-
-    :param inasafe_field_key: Key for the field.
-    :type inasafe_field_key: str
-
-    :param value: Value of the inasafe_default_value.
-    :type value: float, int
-
-    """
-    key = 'inasafe/default_value/%s/%s' % (category, inasafe_field_key)
-    qsetting.setValue(key, value)
-
-
-def get_inasafe_default_value_qsetting(
-        qsetting, category, inasafe_field_key):
-    """Helper method to get the inasafe default value from qsetting.
-
-    :param qsetting: QSetting
-    :type qsetting: QSetting
-
-    :param category: Category of the default value. It can be global or
-        recent. Global means the global setting for default value. Recent
-        means the last set custom for default value from the user.
-    :type category: str
-
-    :param inasafe_field_key: Key for the field.
-    :type inasafe_field_key: str
-
-    :returns: Value of the inasafe_default_value.
-    :rtype: float
-    """
-    key = 'inasafe/default_value/%s/%s' % (category, inasafe_field_key)
-    default_value = qsetting.value(key)
-    if default_value is None:
-        if category == GLOBAL:
-            # If empty for global setting, use default one.
-            inasafe_field = definition(inasafe_field_key)
-            default_value = inasafe_field.get('default_value', {})
-            return default_value.get('default_value', zero_default_value)
-
-        return zero_default_value
-    try:
-        return float(default_value)
-    except ValueError:
-        return zero_default_value
